refactor: Change service layer response type
- Serivce layer에서 반환하는 클래스와 Controller에서 응답하는 클래스 분리
- Controller 응답 타입 ResponseEntity로 통일
- DTO 디렉토리 위치 변경
