+++ src/main/java/com/naekang/wealgo/domain/auth/controller/response/LoginResponseDTO.java
@@ -1,0 +1,14 @@
+package com.naekang.wealgo.domain.auth.controller.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class LoginResponseDTO {
+
+    private String jwtToken;
+
+}
